Refactor tests to match profile names
